refactor: Step1 towards handling circular dependencies
