+++ examples/movie_domain/client/_api/_core.py
@@ -6,7 +6,7 @@ from cognite.client import CogniteClient
 from cognite.client import data_modeling as dm
 from cognite.client._constants import INSTANCES_LIST_LIMIT_DEFAULT
 
-from movie_domain.client.data_classes.core import T_TypeApplyNode, T_TypeNode, T_TypeNodeList
+from movie_domain.client.data_classes._core import T_TypeApplyNode, T_TypeNode, T_TypeNodeList
 
 
 class TypeAPI(Generic[T_TypeNode, T_TypeApplyNode, T_TypeNodeList]):
