+++ examples/movie_domain/client/data_classes/__init__.py
@@ -1,12 +1,10 @@
-from .actors import Actor, ActorApply, ActorList
-from .best_directors import BestDirector, BestDirectorApply, BestDirectorList
-from .ids import RoleId
-from .persons import Person, PersonApply, PersonList
+from ._actors import Actor, ActorApply, ActorList
+from ._best_directors import BestDirector, BestDirectorApply, BestDirectorList
+from ._persons import Person, PersonApply, PersonList
 
 __all__ = [
     "Person",
     "PersonApply",
-    "RoleId",
     "PersonList",
     "Actor",
     "ActorApply",
