+++ examples/movie_domain/client/data_classes/_movies.py
@@ -5,7 +5,7 @@ from typing import ClassVar, Optional
 from cognite.client import data_modeling as dm
 from pydantic import Field
 
-from .core import CircularModelApply, DomainModel, TypeList
+from ._core import CircularModelApply, DomainModel, TypeList
 
 __all__ = ["Movie", "MovieApply", "MovieList"]
