--- examples/movie_domain/client/data_classes/ids.py
@@ -1,7 +1,0 @@
-from __future__ import annotations
-
-from movie_domain.client.data_classes.core import Identifier
-
-
-class RoleId(Identifier):
-    ...
